fix: ensure player is stopped before starting
scenario : we want to play multiplehaptics in succession. Before this commit, the player could still be running while we want to play another haptic. And this causes the player to handle weirdly (skipping/blending pattern, etc.). This commit ensures that before starting a pattern we cancel all the previous ones.